checkqueue: set MAX_SCRIPTCHECK_THREADS to nCores - 1
A fixed MAX_SCRIPTCHECK_THREADS value is not flexible for users to leverage their
cpu resources, and a value large than nCores - 1 doesn't make sense since it only
adds some context switch overhead. Set it to nCores - 1.
Assumption: A user who sets the number of script verification workers is aware of
how this affects the system performance, otherwise he/she leaves it as default
(which is 0)

Signed-off-by: Hao Xu <hao.xu@linux.dev>